Update all price-related fields to use `u128` for greater flexibility, enhance order types with a `user_id` field, and introduce the `Hash32` structure for user identification.

+++ src/execution/match_result.rs
@@ -55,11 +55,11 @@ impl MatchResult {
     }
 
     /// Get the total value executed
-    pub fn executed_value(&self) -> u64 {
+    pub fn executed_value(&self) -> u128 {
         self.transactions
             .as_vec()
             .iter()
-            .map(|t| t.price * t.quantity)
+            .map(|t| t.price * (t.quantity as u128))
             .sum()
     }
